Better cost calculation
The room w/ the charger has a coverage plan that starts at the charger location but ends in some other point. This could potentially be really far away. When calculating paths between rooms, we must always assume the robot, coming from the charger room, starts at the charger location. And if we're returning to the charger room we must not consider the cost of cleaning the charger room because we're only returning to recharge, not clean. This means that the charger room is only cleaned after all other rooms.