Ummm...
The TSP solution I wrote was wrong: 1) Asymmetrical TSP is a very complicated research area that I don't have time to look at, 2) the algorithm was not solving for TSP, but something else. However, thanks to Jacob's bin packing method we don't have to start from scratch! The bin packing works really well, and with a little adjustment to the battery loss level it actually generates more than one bin. All that's left to do is improve the paths. I think either a) force the coverage server to use an inflated map or, b) change the local planner params to accept a pose when any part of the robot's footprint overlaps (makes sense because any part of a roomba can sweep).